requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0

# Resume ingestion
pymupdf>=1.24.0      # PDF extraction
python-docx>=1.1.0   # DOCX extraction
